+++ src/main/java/com/coupong/order/dto/OrderDto.java
@@ -12,11 +12,11 @@ public class OrderDto {
 
     private List<OrderItemDto> orderItems;
 
-    private Integer issuedCouponId;
+    private Long issuedCouponId;
 
     public OrderDto() {}
 
-    public OrderDto(String phoneNumber, String address, List<OrderItemDto> orderItems, Integer issuedCouponId) {
+    public OrderDto(String phoneNumber, String address, List<OrderItemDto> orderItems, Long issuedCouponId) {
         this.phoneNumber = phoneNumber;
         this.address = address;
         this.orderItems = orderItems;
@@ -35,7 +35,7 @@ public class OrderDto {
         return orderItems;
     }
 
-    public Integer getIssuedCouponId() {
+    public Long getIssuedCouponId() {
         return issuedCouponId;
     }
 }
